Add non-Spanish spelling words
The Spanish words that are being wrongly flagged as misspellings
should instead be covered by using both English and Spanish
spell-checking dictionaries.

Some words are not Spanish but I still didn't add them, because it
is likely they will be removed from the code in the near future, or
because I am not certain we really consider them correctly spelled.