Customize dev container git prompt and diff colors
This addresses a few aspects of the dev container's user
experience, though we might revisit them and do something even
better in the future:

1. The default bash prompt for dev containers doesn't show new and
   changed files. It can be made to show unstaged changes by
   setting devcontainers-theme.show-dirty to 1 in the git
   configuration, but this still does not show changes that have
   been staged but not committed.

2. The default bash prompt for dev containers doesn't show when we
   are behind and/or ahead of our remote.

3. The green color used in diffs is hard to distinguish from
   surrounding text in some light terminal themes, including the
   default light them used in a codespace.

(1) and (2) are improved by using a slightly customized classic
(__git_ps1 based) git prompt. In the future, if we want, we could
get the advantage of both kinds of prompts by doing more detailed
customizations of one of them to incorporate what we consider good
about the other.

(3) is improved by changing that green color to blue, which is easy
to distinguish from surrounding text in both light and dark themes.