Download and run Mambaforge installer in /tmp
Although it's possible for /tmp to be on storge optimized for
short-term use, the real reason to do this is for when the dev
container is used locally but with the workspace on the host.

This is a less common way to use a dev container locally, compared
to (re)cloning the repository in the dev container (in which case
the workspace is stored on a volume associated with the container).
In this situation, there are a few advantages of ensuring that the
workspace directory is not used to hold the Mambaforge installer:

- Accessing files on the host from the dev container may be slower.
  This is rarely a problem on GNU/Linux hosts, but it is always a
  problem on Windows and macOS hosts (which is why the usual
  approach of cloning the repo in the container is better on those
  systems... but it is not always done).

- If the installation fails, Mambaforge-Linux-x86_64.sh could be
  left over in the workspace directory, which could be annoying or
  confusing or even, occasionally, lead to it being accidentally
  committed.

- It is possible, though unlikely, that something about the host
  filesystem is unsuitable. For example, it might not permit files
  to be marked executable, or it might not have enough free space.

None of these are really big problems, but the fix is simple.