Show the shallow structure more readably
Note that, as before, the structure is "shallow" only in the sense
that instead of showing the deepest parts -- the contents of the
lists -- it shows "..." in their place.

This no longer uses cheap_repr.

Instead, it round-trips the data by encoding to JSON and decoding
it, which has the effect of recursively converting the tree of
Python objects, some of whose types are non-builtins defined in the
openai module, to a tree of Python objects each of which is a
builtin.

With all nodes involved being builtins, pprint knows how to format
it readably, and calling pprint.pp with an appropriate depth is
sufficient.