Try to more clearly indicate when CI fails from no API key
This attempts to have a single job that just checks for the API key
and fails if it is absent, with jobs that need the API depending on
that job and thus not running at all if it fails.

I'm not sure how well this will work, because whether or not the
API key is available may depend on the event trigger. But in the
common case where it is available with the push event but
unavailable in the pull_request event, those are events are on
different repositories, so maybe doing it this way is sufficient.