Have tests cache embedding calls if requested
Currently the user can request this using the
TESTS_CACHE_EMBEDDING_CALLS environment variable.

This actually only moderately decreases the number of API calls,
with 16 cache misses and only 12 cache hits. That is, over 57% of
the API calls are still being made. As such, the complexity of
doing this might not be justified. It might also make sense to look
into simplifying the code, or adding one or more dev dependencies
to allow some code, at least code in _test_helpers.py, to be
removed.

This commit contains some debugging cruft, to be deleted soon.