Cache embedding calls within each CI test job
This still does not cache them *across* jobs, it just enables the
recently added caching functionality in test_embed.py for each job
separately.

(Caching results across jobs would be undesirable, because it
wouldn't test that each combination of platform and Python version
is able to get embeddings. So it would be a more complicated way to
achieve almost the same goal as removing the jobs.)