Refactor cache logging; report total hits/misses
Total hits and misses are logged at the INFO level (and so are
shown at DEBUG or INFO level), while reports of each individual hit
and miss only appear at DEBUG level.

If no caching happens, then no hit/miss is shown. That way, when
caching has not been enabled, no unnecessary message appears.

This is all just for testing. At this time, causing tests to make
fewer calls to the OpenAI embeddings API endpoint is the only use
of caching in this project.