Limit CI test parallelism (#2)
To make API requests at a slower rate.

If possible, I think it's best if users of the OpenAI API who are
still in their unpaid trial period can still try out this project.
Such users have lower rate limits, as detailed in:

https://platform.openai.com/docs/guides/rate-limits/what-are-the-rate-limits-for-our-api

The change made here still allows jobs to run in parallel, but the
step that actually runs the tests can only run in one job at a time
(per repository).

The GitHub Action used to achieve this mutual exclusion (treating
the step that runs the unit tests as a critical section) uses a
branch of the repository, created and destroyed automatically as
necessary, to synchronize access. (It takes advantage of how Git
synchronizes push operations.) But because the number of contending
jobs may sometimes be high, it is possible for a
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/429 error
to occur on the repository URL. If this happens, especially if it
happens with a *different* action such as actions/checkout on
another job, a failed job may result, and the failed job need not
be one of the jobs that is waiting to enter the critical section.
Although the problem could be worked around by re-running affected
workflows, it would be annoying and confusing. So if it happens
(and more than very rarely), then it may be best to revert this.

This squashes some commits that include false starts and
troubleshooting.

See https://github.com/EliahKagan/EmbeddingScratchwork/pull/2 for
full details on all original commits that made it into this
squashed commit (beyond their messages, all listed below).

---

* Decrease CI test matrix parallelism

To make API requests at a slower rate.

If possible, I think it's best if users of the OpenAI API who are
still in their unpaid trial period can still try out this project.
Such users have lower rate limits, as detailed in:

https://platform.openai.com/docs/guides/rate-limits/what-are-the-rate-limits-for-our-api

* Let CI test setup steps run with full concurrency

The approach in the previous commit of limiting jobs took way too
long because most of the time was provisioning micromamba. Caching
to GitHub Actions storage wouldn't be a very good solution, because
the cached data would be different (and thus shouldn't be shared)
for each of the fifteen jobs, and because we would want a very
short cache window (with conda, at least currently, we are not
pinning specific dependency versions, and CI should reflect that).

This new approach attempts to lock just the "Run Unit Tests" step
using a mutex. It should let just one of these steps run at a time,
across all jobs, while still allowing each job to provision with
micromamba concurrently. However, this may not be worth it, even if
it works, because of the complexity entailed in using an additional
branch of the repository to maintain the mutex state, as the mutex
action being used does.

* See if the mutex action can be run without Docker

So it works on macOS and Windows runners as well as GNU/Linux.

* Try another mutex action that doesn't use Docker

* Switch to my fork of the mutex action

* Switch to v1 temporarily to investigate errors

* Bisect to find breakage (step 1)

This specifies a commit hash before most changes.

* Bisect to find breakage (step 2)

This specifies the hash referring to the same commit as the v1 tag.

* Bisect to find breakage (step 3a)

* Bisect to find breakage (step 3b)

This is the same hash as the tip of the main branch.

* Switch to (now available) v2 tag