Validate blocking configuration (#5)
* Validate blocking configuration

* Change validation job to record the setting

This lets "Set up mutex" just test the value as interpreted.

This also shows a message if the value is unset (empty).

* Debug use-mutex configuration

* Set the output properly (I hope)

* Use "0" and "1" for use-mutex

Since "false" didn't evaluate as false.

* Put back to "true" and "false"

"1" and "0" were no better.

The problem still has to be fixed, of course.

* Compare to the string "false"

Since just evaluating the string as a condition didn't work.

* Fix config variable doc URL

For a more informative message.

+ Tweak message wording.

* Remove debugging cruft

* Check effect of ::warning

* Try backslash escapes in ::warning string

* Try literal newline

* Try URL-encoded newline

I've tried \n and a literal newline. This tries %0A.

* Newlines in step, spaces in annotation

As I don't think I can get newlines in the annotation.

* Convert from if-elif with grep to case

* Strip trailing whitespace

The whitespace was accidentally introduced in the previous commit.