Make blocking config retrieval code its own script (#7)
* Make blocking config retrieval code its own script

It was taking up more space than anything else in test.yml, and
while important, it is not fundamental to what test.yml is actually
doing (setting up to run the automated tests and running them).
This should make that workflow file easier to read and understand.

[This also tweaks the style of the bash code in the script, to
prefer single over double quotes and to prefer printf over echo.]

* Fix path to get-blocking-config script

* Try again to fix get-blocking-config script path

* It helps if I check out the repo first

This problem was not related to paths.

Since I have to check out the repository, the space and code
complexity savings in test.yml are not as pronounced. But it is
still shorter and simpler. Maybe more important is how having the
logic in a script lets shellcheck scan it.