Make editable installs in CI workflows
This modifies the two workflows that provision with micromamba,
adding a "Make editable install" step that runs "pip install -e .".
This currently shouldn't make a difference, but the CI workflows
shouldn't unnecessarily install the project in a different way
from the way it is otherwise being installed. (In particular, in
test.yml, if doing this step somehow ever causes a test to fail, we
will want CI to discover that.)