State model name in the logo alt and title text
This changes only the readme, not the logo itself.

The reason for this change is to avoid misleading readers into
thinking that embeddings, and their similarities, are properties
of texts by themselves. (They are, of course, properties of texts
taken together with an embedding model, and vary across changes of
model.)