Document how to enable workflows (#32)
* Document how to enable workflows

My previous wording had effectively suggested that the inherited CI
workflows would be enabled automatically in a fork, which is not
the case. This commit addresses how to enable them in a a forked
repository's Actions tab.

I don't want to put an enormous amount of what is effectively
general GitHub documentation in the readme, since the readme is
already verging on being excessively long. But this is
unfortunately not well documented in the GitHub docs. (See
https://github.com/github/docs/issues/15761, which was closed
as completed, but appears actually to have been declined.)

In the Actions tab of a newly created fork, this message is shown:

> Workflows aren’t being run on this forked repository
>
> Because this repository contained workflow files when it was
> forked, we have disabled them from running on this fork. Make
> sure you understand the configured workflows and their expected
> usage before enabling Actions on this repository.

There is a button "I understand my workflows, go ahead and enable
them" as well as a link "View the workflows directory".

I've largely not documented the details of this; I note them in
this commit message so it's clear what the changes in this commit
refer to in case the GitHub user interface changes in the future,
and for insight.

Instead, I've linked to two pages where they are discussed: the
GitHub docs bug report, and documentation for an unrelated project.
These sources are imperfect, since neither is intended as *general*
documentation, and maybe the information should be reproduced in
our readme, but for now I think this may be enough.

* Fix line length in README

---------

Co-authored-by: David Vassallo <vassallo.davidm@gmail.com>