Add a couple of our formatting settings for VS Code
The important change is making it so VS Code uses LF line endings,
since that's the convention we're using for the project. This is
set for Git in .gitattributes, but VS Code on some platforms,
including the web editor, doesn't automatically respect that.

This also adds the PEP-8 recommended 79-character line length to
the rewrap extension's settings, so Alt+Q does that without
requiring it to be otherwise configured.