Rename one of the "response" local variables
The "response" from calling openai.Embedding.create is different
from the response returned by requests.post: the former corresponds
to the result of calling the json method on the latter. So this
renames the "response" from openai.Embedding.create, to avoid
confusion when the functions are read together or compared.

This commit also makes docstring style a little more consistent.