Use the features of OpenAI's dict subclass
This refactors embed_one and embed_many so they take advantage of
the more specific typing (than dict) that we get from the OpenAI
Python library. This is less error prone, at least in theory, in
that static tools (in an editor or otherwise) may be more able to
identify an incorrectly typed attribute name than an incorrectly
typed dict key name. A possibly more significant benefit is
readability (the code is neater, using less punctuation).

This is one of two approaches I would consider reasonable in
further refactoring that code. The other approach is to continue
subscripting but factor out the identical (or very similar) code
between embed_one and embed_one_req and between embed_many and
embed_many_req. This code becomes less similar as a result of the
approach I've taken here, so both refactorings cannot be done.

The reason I've picked the former refactoring is that it is
consistent with (and even serves) the goal of clearly demonstrating
how to use the three major techniques for retrieving OpenAI
embeddings, allowing their similarities and differences, and their
advantages and disadvantages, to be noticed.