Let the test workflow be manually triggered
As well as still running on the push and pull_request events.

This adds the workflow_dispatch trigger so that it is not merely
possible to re-run an existing run (which already works), but also
to trigger a new workflow run, not dependent on or specifically
associated with any particular automatic run, by going to the
GitHub repository's Actions tab, clicking the workflow name on the
left side, and then clicking "Run workflow".

"Run workflow" appears for workflows that list the
workflow_dispatch event as a trigger. For more information about
the user interface, see:

https://docs.github.com/en/actions/managing-workflow-runs/manually-running-a-workflow?tool=webui