Decrease request volume and thread stack size
This supersedes b363281. The problem was the broken regex (fixed in
0df6fbf and 386584f). There's no need to send that many requests to
trigger (and test) rate limiting; the reduced number here, given
enough concurrency, appears sufficient.

Of course, this many requests should ordinarily be (a) avoided, if
possible, or (b) done with asynchronous programming or some other
technique that avoids creating hundreds of threads. But here I am
doing automated testing of synchronous code.