Use consistent quoting style across whole project
Some code preferred single quotes and other code preferred double
quotes. I've changed the style to use single quotes everywhere
reasonably applicable.

I picked this default not out of the belief that single quotes
ought to be the default (I have no opinion), but because most of
our *newer* code, including code that is in feature branches as of
this writing, uses single quotes. We might decide to change it to
double quotes later, but this makes it consistent, which I think is
more important than which convention we choose.

Note that this only applies to Python code. This applies to both
modules and notebooks. But it doesn't apply to other languages
(bash, yaml, toml, json), which do not have the same symmetry of
meaning of "" vs. '' that Python has.