Fix regex and increase iteration count
Now that I'm using log_context.output, it *does* include the prefix
from the logger, which I actually do want to test, but I had
previously removed it from the regex because the old way I was
checking the log didn't include it.

For some reason, the thread and iteration counts, which did not
change in the reorganization done in the previous commits, now
see insufficient to trigger rate limiting. I suspect this is due to
variable delays associated with CPython implementation details,
arising from the different code constructs from the reorganization
(such as using a derived class for the thread). If so, then it was
pretty close to the threshold anyway, so this change may be good.

However, this does point to the brittle nature of this test. I'm
unsure if test_backoff.py should actually be retained, once the
forthcoming refactoring of backoff in the requests versions is
completed.