Use concurrent.futures.ThreadPoolExecutor
For the backoff tests, instead of raw threads.

The most important benefit in this case is that it takes care of
raising exceptions from worker threads in the calling thread,
which I was failing to make happen.

However, this is also just a better practice in general, decreases
how much code is needed, and more clearly expresses the goal of
running a specific number of jobs on a specific number of threads.