Do closer to the original approach (but executor)
Instead of expressing all 4000+ jobs separately and trying to have
the executor chunk the work, this makes _THREAD_COUNT jobs, but
still uses the executor (with _THREAD_COUNT threads).

Unlike previous executor approaches I've tried, this seems to work
(for the weird use case I here mean--executors generally work just
fine). Assuming that is the case, this remains preferable to using
"raw" threads because (a) I avoid having to code in terms of the
individual threads, and more importantly (b) exception raised in
worker threads will propagate to the caller, preventing erroneous
test passes that ought to be failures or errors, due to unhandled
exceptions from the code under test.