Make data_dir kw-only arg somewhat more robust
This gives data_dir a default value of None, which is replaced
with DEFAULT_DATA_DIR when called. That way, there is a constant
for calling code to reference, and monkey-patching the constant
(though that should usually be avoided) works properly.

Besides None, data_dir may (still) be a path-like object or a str.

This is still not as robust as it should be. Because _disk_cache's
wrapper is used both for embed_one* and for embed_many* functions
and its first parameter is named text_or_texts, calls of this form
continue to fail with TypeError:

    embed.cached.embed_one(text='hello')
    embed.cached.embed_many(texts=['hello', 'goodbye'])

That is, the wrapper function cannot be called with a keyword
argument using original parameter name.

Furthermore, calling help or inspect.signature on such a wrapper
shows the original signature, even though that cannot be used.

Adding a data_dir parameter compounds both of those problems:

- When a function can never be called with multiple arguments, it
  is nearly always sufficient to pass the argument positionally.
  But now that functions like embed.cached.embed_one accept two
  arguments, people will sometimes want to pass both as keyword
  arguments.

- help and inspect.signature on the wrapper don't show data_dir.

It is practical to solve these problems, but we might not want to
bother doing so until we've decided if the current interface and
code organization is what we want. In contrast, adding a data_dir
argument increases testability of the existing approach (which may
also make it easier to change the code).