Correct and simplify the test helper module's docstring (#53)
tests._helpers.__all__ currently listes two functions, neither of
which is a decorator, and only one of which is at all related to
decorators. The docstring's wording was from an earlier design I
considered for the _helpers module, and I never fixed it.

I've also taken this opportunity to remove the mention of how it is
used in test_embed.py. That's true, but it will become misleading
as soon as any other test module uses the _helpers module.