Use the same boolean parsing rules everywhere (#56)
This modifies tests/_helpers.py to parse the values of environment
variables that represent booleans according to the same rules used
in .github/workflows/get-blocking-config. There is no design reason
for them to behave differently, I just allowed them to diverge
according to what was easy to code in Bash vs. Python. This fixes
that.

I think it's likely we'll want to make _helpers._getenv_bool public
in the future (renaming it accordingly), for use in future tests.
But nothing would use that currently, so I've kept it nonpublic for
now.