Fix expected path construction on Windows (#65)
The code under test builds paths with the operating system's
default path separator (as it should), but the test code contained
a literal "/" separator. On Windows, a "\" appears instead, failing
the test. This fixes that by using pathlib.Path in the test logic.
I've done it a little differently from the code under test, though
it may be worth exploring ways to further reduce similarity between
the test and the code under test.