Fix decorator placement in embed_many* tests
I had wrongly put on `@maybe_cache_embeddings_in_memory` on
TestEmbedMany, which supplied (un)patching logic to all its test
case methods, but the code that needed to be wrapped to carry out
monkey-patching was instead the setUp method.

This fixes that, rewords the documentation to make this bug less
likely (by avoiding implying that when the decorator can only
reasonably be applied to methods that represent test cases), and
removes the working notes I made in investigation/ on the problem.

This succeeds at restoring the number of in-memory cache misses and
hits from before the overhaul of the tests' optional in-memory
caching logic.

The decorator can of course also be applied to the class, or to
specific test-case methods in the class, if that test class later
calls embedding functions outside its setUp fixture.