Support in-memory caching in on-disk caching tests
Now that the caching logic is presented in a way where the two
forms of caching (on disk, as part of the code under test, and
in-memory, as an optional enhancement to the tests themselves) are
not likely to be confused, and now that the tests' optional
in-memory caching is overhauled to use monkey-patching so it
applies to indirect as well direct calls to the embed.embed_*
functions, it is now both reasonable and possible to mix them.

This commit extends the tests' optional in-memory caching of
embed.embed_* results to the tests of embed.cached.embed_*
functions.

Note that calls to disk caching functions (embed.cached.embed_*)
are never under any circumstances bypassed by this optional
in-memory caching done as part of the tests. Instead, some calls to
the embedding functions that do *not* cache to disk (embed.embed_*)
are bypassed both in their own tests and in embed.cached.embed_*
tests (since the embed.cached.embed_* functions themselves call the
embed.embed_* functions).

Of course, optional in-memory caching remains solely a feature of
the test suite, and remains only done at all if enabled (and
remains turned off by default, but our CI workflows turn it on).

Before this change, when the tests' optional in-memory caching was
turned on, there were 24 misses and 18 hits. Now, when it is turned
on, there are 27 misses and 21 hits. This is to say that, so far,
the embed.cached.embed_* tests are actually benefiting (in the
sense of percentage reduction in API calls) *more* than the
embed.embed_* tests: the embed.cached.embed_* tests made 6 calls to
embed.embed_* functions, and now, when the tests do in-memory
caching, make 3 calls, with a full 50% savings. This amount of
benefit may or may not continue as the embed.cached.embed_* tests
are extended (those tests are not done being written yet).