Improve in-memory cache stats display + docstrings
Among other improvements, this fixes a bug where the object
representing the stats was always truthy (but evaluated as a
boolean to decide whether to print a message). I had introduced
that bug when I changed the stats from dict to SimpleNamespace. Now
it is an attrs data class with with a custom __bool__.

This also adds docstrings to Caller. (This is unrelated, in terms
of function, but I was introducing a class for the stats, and it
was convenient to write method docstrings for both at the same
time.)

Some of these changes facilitated reordering the top-level code in
tests._helpers, which is also done here.