Parameterize the class in test_cache by func
Instead of parameterizing each method the same way.

I think this is equally clear and intuitive. I renamed a method
that otherwise might make the meaning of self.func ambiguous in it.

I've also taken this opportunity to change the class name to
reflect its current scope of purpose (it can be changed again
later, of course) as well as to avoid any confusion that may arise
between the two kinds and purposes of caching in the tests.