Have test_cached refer to embed.cached as cached
Although arguably inelegant, this "from" import has the benefit of
making the parameterization code considerably more readable.

The import now used, "from embed import cached", has the additional
effect -- like "import embed.cached" would have but unlike the
import being replaced (just "import embed) -- of not relying on the
"embed" package importing its "cached" submodule. That might be
regarded as good or bad, depending on whether we want something in
the tests to fail if we don't do this. I'm not actually sure will
end up continuing to do it, and I don't consider verifying it to be
a current goal of the tests.