Have embed.cached use its own logger
Instead of the root logger.

This is a good practice, but there is a more specific reason to do
it now: we cannot really guarantee that tests that the only message
from the root logger or any of its descendants (i.e., any logger)
with INFO or higher severity is the message reporting the on-disk
cache hit or miss.