Have test modules use their own loggers
And have the remaining unscoped (i.e., root logger scoped)
assertLogs assertion scope to a specific logger (which is actually
one from a library, rather than one of the loggers in our code).

This change corresponds *roughly* to 6a36ee6 (the previous commit),
but for logging in tests rather than logging in code under test.