Make run-test-scripts an sh script
It doesn't use any Bash features, so it may as well specify /bin/sh
as the interpreter. (This is what we're doing with other shell
scripts in the project that rely only on POSIX shell features, such
as .devcontainer/onCreate.)

The script is otherwise the same as in 74e61c2.