Make Caller take a supplier again
When I had tried this out before, this approach seemed to make the
code using Caller too complicated and hard to read. That no longer
appears to be the case. Due in part to other changes between then
and now, as well as differences in the style I used this time, this
now seems clearer as to its intentions at the call site, and also
far more obvious *how* it works and *that* it works.

Furthermore, this approach eliminates the brittleness of guessing
the module and checking the guess in a half-baked and confusing
way. Now that Caller does less, it's shorter and simpler.