Suppress pylint consider-using-f-string
This turns off "C0209: Formatting a regular string which could be a
f-string" project-wide.

In general, that is a useful warning, particularly for older
projects that have code from before f-strings were available in all
supported Python versions, or written by authors not accustomed to
f-strings. But we are unlikely ever to use .format or % for string
formatting without first considering f-strings. So the likelihood
of true positives from this, in this project, at least at present,
is low.

I'm adding this suppression now because the diagnostic is raised on
the use of .format added in the preceding commit (11e35f4), which I
do in the tests so that all expected log messages' forms are
readily clear and presented in the same way so as to be comparable
to one another.