Work around version incompatibilities
The pylint-actions plugin only supports Python 3.9 and later. This
is not a problem, given that we are only ever actually using its
functionality with Python 3.11. However, this keeps us from listing
the package for it in environment.yml, because although that file
*also* lists Python 3.11, it is automatically modified for other
Python versions in CI test jobs (which are not related to the CI
pylint job that uses the plugin). So instead, we install the plugin
only when it is needed (i.e., for the CI pylint job). But pylint
understandably complains if it's configured to use a plugin that
isn't present, and it would be bad to diminish the experience of
using pylint outside CI. So this has that job edit pyproject.toml
to add the configuration to use the plugin that is only installed
in that same CI job. (The edit is of course never committed.)