Fix Python 3.7 syntax error; improve helper name
We're not ever actually *calling* anything in tests/_audit.py on
Python 3.7, because the auditing event we're interested in (as well
as sys.addaudithook) were only added in Python 3.8, and the tests
are conditionally skipped for 3.7. However, because the := is a
syntax error prior to its introduction in 3.8, this will cause an
exception to be raised when the module is imported. So I've
refactored to remove the use of := (and improved the name of the
helper function while I'm at it).