Remove subscribe/unsubscribe locking
Locking would be needed in general but not the way we are using, or
are ever likely to use, _audit in our tests.

If the code in _audit is ever extracted into a *separate* project
and put on PyPI, then the locking should probably be added back in.

(It is a non-public helper module of our tests, so the absence of
locking does not block future packaging/publishing of *this*
project on PyPI.)